Move scenario tests under mw-dev harness
