+++ .github/workflows/scenario-dev.yml
@@ -29,7 +29,7 @@ jobs:
     if: ${{ github.event_name != 'workflow_dispatch' || !inputs.skip_slow }}
     uses: ./.github/workflows/_image-build.yml
     with:
-      dockerfile: tests/scenario/Dockerfile
+      dockerfile: tests/mw-dev/scenario/Dockerfile
       image-name: duckgres
       tag: scenario-runner-${{ github.run_id }}-${{ github.run_attempt }}-arm64
       platform: linux/arm64
@@ -85,7 +85,7 @@ jobs:
           go-version-file: go.mod
 
       - name: Test scenario workflow scripts
-        run: go test -count=1 ./tests/scenario ./tests/mw-dev
+        run: go test -count=1 ./tests/mw-dev/scenario ./tests/mw-dev
 
       - name: Configure AWS credentials (OIDC)
         uses: aws-actions/configure-aws-credentials@61815dcd50bd041e203e49132bacad1fd04d2708 # v5.1.1
