+++ tests/mw-dev/scenario/core/artifacts.go
