+++ tests/mw-dev/scenario/core/artifacts_test.go
