+++ tests/mw-dev/scenario/core/catalog.go
