+++ tests/mw-dev/scenario/core/catalog_test.go
