+++ tests/mw-dev/scenario/core/templates.go
