+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/dbt_project.yml
