+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/facts/fct_activation_funnel.sql
