+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/facts/fct_sessions.sql
