+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/intermediate/int_event_days.sql
