+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/intermediate/int_person_first_seen.sql
