+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/intermediate/int_sessionized_events.sql
