+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/marts/mart_product_kpis_daily.sql
