+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/schema.yml
