+++ tests/mw-dev/scenario/dbt/posthog_frozen_project/models/staging/stg_events.sql
