+++ tests/mw-dev/scenario/scenarios/posthog_frozen_dbt.yaml
