+++ tests/mw-dev/scenario/scenarios/posthog_frozen_metadata.yaml
