+++ tests/mw-dev/scenario/scenarios/provision_smoke.yaml
