+++ tests/mw-dev/scenario/sql/errors.go
