+++ tests/mw-dev/scenario/sql/setup_frozen_views.sql
