+++ tests/mw-dev/scenario/sql/steps.go
@@ -10,8 +10,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/posthog/duckgres/tests/scenario/core"
-	"github.com/posthog/duckgres/tests/scenario/provision"
+	"github.com/posthog/duckgres/tests/mw-dev/scenario/core"
+	"github.com/posthog/duckgres/tests/mw-dev/scenario/provision"
 	"gopkg.in/yaml.v3"
 )
